+++ src/backend/api/managers/auth_manager.py
