+++ src/backend/api/managers/user_manager.py
