+++ test/backend/base.py
@@ -1,6 +1,6 @@
 from flask_testing import TestCase
 
-from src.backend import db
+from api import db
 from manage import app
 
 
@@ -8,7 +8,7 @@ class BaseTestCase(TestCase):
     """ Base Tests """
 
     def create_app(self):
-        app.config.from_object('src.backend.config.TestingConfig')
+        app.config.from_object('api.config.TestingConfig')
         return app
 
     def setUp(self):
